- refactor app  - move helper functions to separate files
