+++ src/CarDetails/index.jsx
@@ -4,6 +4,7 @@ import axios from 'axios';
 import { v4 as uniqueKey } from 'uuid';
 import Styled from "./styles";
 import Spinner from "../Spinner";
+import {getCarDetails} from "./helpers";
 
 const apiKey = '4eb5431794042791adec007a12498a980838beb0aaf0994ed086a1edd44b8d5c';
 let subscription;
@@ -19,15 +20,17 @@ function CarDetails() {
       setContentInProgress(true);
       axios.get(`https://api.overheid.io/voertuiggegevens/${licenseNumber}?ovio-api-key=${apiKey}`)
         .then((response) => {
-          setViewData(getFinalData(response.data));
-          dataLoaded(setContentInProgress, setNoData);
+          setViewData(() => getCarDetails(response.data));
           setErrorMessage(false);
           ObserverService.emitImagesUpdate();
         })
         .catch((error) => {
           setErrorMessage(error.response?.data?.error);
-          dataLoaded(setContentInProgress, setNoData);
           setViewData(false);
+        })
+        .finally(() => {
+          setContentInProgress(false);
+          setNoData(false);
         });
     });
     return () => subscription.unsubscribe();
@@ -53,19 +56,3 @@ function CarDetails() {
 }
 
 export default CarDetails;
-
-function getFinalData(apiData) {
-  const tradeName = apiData.merk + ' ' + apiData.handelsbenaming;
-  const date = apiData.datum_eerste_toelating;
-  const fuelType = apiData.brandstof[0]?.brandstof_omschrijving;
-  const labels = ['Trade name', 'Date of first admission', 'Fuel description'];
-  const values = [tradeName, date, fuelType];
-  return labels.reduce(
-    (prev, curr, index) => ([...prev, { id: uniqueKey(), label: curr, value: values[index] }]),
-    []);
-}
-
-function dataLoaded(setContentInProgress, setNoData) {
-  setContentInProgress(false);
-  setNoData(false);
-}
