+++ src/RandomImages/consts.js
@@ -1,0 +1,18 @@
+export const BREAKPOINTS_CONFIG = {
+  315: {
+    slidesPerView: 1,
+    spaceBetween: 10,
+  },
+  482: {
+    slidesPerView: 1,
+    spaceBetween: 10,
+  },
+  600: {
+    slidesPerView: 2,
+    spaceBetween: 10,
+  },
+  1200: {
+    slidesPerView: 3,
+    spaceBetween: 30
+  },
+};
